requirements.txt
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
graphql-core==3.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.5.1
MarkupSafe==2.1.3
python-dotenv==1.0.0
python-http-client==3.3.7
requests==2.31.0
sendgrid==6.10.0
sgqlc==16.3
soupsieve==2.5
starkbank-ecdsa==2.2.0
types-requests==2.31.0.10
urllib3==2.1.0
waitress==2.1.2
Werkzeug==3.0.1
